Reorg to add more wms methods
