+++ xpublish_wms/wms/get_map.py
@@ -23,7 +23,10 @@ from xpublish_wms.utils import to_lnglat
 logger = logging.getLogger(__name__)
 
 
-class OgcWms:
+class GetMap:
+    '''
+    TODO - Add docstring
+    '''
     TIME_CF_NAME: str = "time"
     ELEVATION_CF_NAME: str = "vertical"
     DEFAULT_CRS: str = "EPSG:3857"
